Adjust new changes in yiisoft/db.
