+++ tests/TestSupport/TestTrait.php
@@ -1,0 +1,134 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Sqlite\Tests\TestSupport;
+
+use ReflectionClass;
+use ReflectionObject;
+use Yiisoft\Db\TestUtility\IsOneOfAssert;
+
+trait TestTrait
+{
+    /**
+     * Asserting two strings equality ignoring line endings.
+     *
+     * @param string $expected
+     * @param string $actual
+     * @param string $message
+     */
+    protected function assertEqualsWithoutLE(string $expected, string $actual, string $message = ''): void
+    {
+        $expected = str_replace("\r\n", "\n", $expected);
+        $actual = str_replace("\r\n", "\n", $actual);
+
+        $this->assertEquals($expected, $actual, $message);
+    }
+
+    /**
+     * Asserts that value is one of expected values.
+     *
+     * @param mixed $actual
+     * @param array $expected
+     * @param string $message
+     */
+    protected function assertIsOneOf($actual, array $expected, $message = ''): void
+    {
+        self::assertThat($actual, new IsOneOfAssert($expected), $message);
+    }
+
+    /**
+     * Gets an inaccessible object property.
+     *
+     * @param object $object
+     * @param string $propertyName
+     * @param bool $revoke whether to make property inaccessible after getting.
+     *
+     * @return mixed
+     */
+    protected function getInaccessibleProperty(object $object, string $propertyName, bool $revoke = true)
+    {
+        $class = new ReflectionClass($object);
+
+        while (!$class->hasProperty($propertyName)) {
+            $class = $class->getParentClass();
+        }
+
+        $property = $class->getProperty($propertyName);
+
+        $property->setAccessible(true);
+
+        $result = $property->getValue($object);
+
+        if ($revoke) {
+            $property->setAccessible(false);
+        }
+
+        return $result;
+    }
+
+    /**
+     * Invokes a inaccessible method.
+     *
+     * @param object $object
+     * @param string $method
+     * @param array $args
+     * @param bool $revoke whether to make method inaccessible after execution.
+     *
+     * @return mixed
+     */
+    protected function invokeMethod(object $object, string $method, array $args = [], bool $revoke = true)
+    {
+        $reflection = new ReflectionObject($object);
+
+        $method = $reflection->getMethod($method);
+
+        $method->setAccessible(true);
+
+        $result = $method->invokeArgs($object, $args);
+
+        if ($revoke) {
+            $method->setAccessible(false);
+        }
+        return $result;
+    }
+
+    /**
+     * Adjust dbms specific escaping.
+     *
+     * @param array|string $sql
+     *
+     * @return string
+     */
+    protected function replaceQuotes($sql): string
+    {
+        return str_replace(['[[', ']]'], '`', $sql);
+    }
+
+    /**
+     * Sets an inaccessible object property to a designated value.
+     *
+     * @param object $object
+     * @param string $propertyName
+     * @param $value
+     * @param bool $revoke whether to make property inaccessible after setting
+     */
+    protected function setInaccessibleProperty(object $object, string $propertyName, $value, bool $revoke = true): void
+    {
+        $class = new ReflectionClass($object);
+
+        while (!$class->hasProperty($propertyName)) {
+            $class = $class->getParentClass();
+        }
+
+        $property = $class->getProperty($propertyName);
+
+        $property->setAccessible(true);
+
+        $property->setValue($object, $value);
+
+        if ($revoke) {
+            $property->setAccessible(false);
+        }
+    }
+}
